+++ houssg/src/components/owner/register/element/StepMover.tsx
@@ -1,74 +1,34 @@
 import styled from 'styled-components';
 import { StepMoverType } from '../../../../types';
 import { color } from '../../../../assets/styles';
-import api from '../../../../api/api';
-import { ownerUrl } from '../../../../assets/constant';
+import { onRegiFunnelData } from '../../../../helper';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { ownerKey, ownerRoute } from '../../../../assets/constant';
+import { useNavigate } from 'react-router-dom';
 
 const StepMover: React.FC<StepMoverType> = ({ goStep, step, data, inactive, last }) => {
-	const onRegiFunnelData = () => {
-		// !inactive
-		const formData = new FormData();
-		if (data.houseImageFile !== null) {
-			formData.append('file', data.houseImageFile as File);
-		}
+	// const { mutate } = useMutation(() => onRegiFunnelData(data));
+	const queryClient = useQueryClient();
+	const navigate = useNavigate();
 
-		const requestData = {
-			accomName: data.name,
-			accomAddress: data.targetAddress,
-			teleNumber: data.houseNumber,
-			accomCategory: data.houseType,
-			accomDetails: data.detailText,
-			checkIn: data.checkIn,
-			checkOut: data.checkOut,
-			businessNumber: data.businessNum,
-			facilityDto: data.houseService,
-		};
+	const { mutate } = useMutation({
+		mutationFn: () => onRegiFunnelData(data),
+		onSuccess: () => {
+			queryClient.invalidateQueries({ queryKey: [ownerKey.myHouseList] });
+			alert('성공');
+			navigate(ownerRoute.management);
+		},
+	});
 
-		console.log(data.houseImageFile, '이미지파일');
-		console.log(requestData, 'request');
-
-		const json = JSON.stringify(requestData);
-		const blob = new Blob([json], { type: 'application/json' });
-		formData.append('request', blob);
-		return api.post(ownerUrl.houseRegister, formData, { headers: { 'Content-Type': 'multipart/form-data' } });
+	const onRegiHandler = () => {
+		mutate();
 	};
-	// if (data.houseService) {
-	// 	const dtoKey = ['nearbySea', 'parkingAvailable', 'pool', 'spa', 'wifi', 'twinBed', 'barbecue', 'noSmoking', 'luggageStorage', 'freeMovieOtt'];
-
-	// 	const resultObject: Record<string, number> = {};
-
-	// 	// 배열의 길이를 기준으로 반복하면서 객체를 만듭니다.
-	// 	for (let i = 0; i < dtoKey.length; i++) {
-	// 		const key = dtoKey[i];
-	// 		const value = data.houseService[i];
-	// 		resultObject[key] = value;
-	// 	}
-	// 	const requestData = {
-	// 		accomName: data.name,
-	// 		accomAddress: data.targetAddress,
-	// 		teleNumber: data.houseNumber,
-	// 		accomCategory: data.houseType,
-	// 		accomDetails: data.detailText,
-	// 		checkIn: data.checkIn,
-	// 		checkOut: data.checkOut,
-	// 		businessNumber: data.businessNum,
-	// 		facilityDto: data.houseService,
-	// 	};
-
-	// 	console.log(data.houseImageFile, '이미지파일');
-	// 	console.log(requestData, 'request');
-
-	// 	const json = JSON.stringify(requestData);
-	// 	const blob = new Blob([json], { type: 'application/json' });
-	// 	formData.append('request', blob);
-	// 	return api.post(ownerUrl.houseRegister, formData, { headers: { 'Content-Type': 'multipart/form-data' } });
-	// }
 
 	return (
 		<StepBtnAligner>
 			{!!step && <StepBtn onClick={() => goStep(step - 1)}>이전</StepBtn>}
 			{last ? (
-				<StepBtn $inactive={inactive} onClick={onRegiFunnelData}>
+				<StepBtn $inactive={inactive} onClick={onRegiHandler}>
 					등록하기
 				</StepBtn>
 			) : (
