+++ houssg/src/hooks/useImageConverter.ts
@@ -7,7 +7,6 @@ export const useImageConverter = () => {
 
 	const setIncodedImg = async () => {
 		if (imgRef.current?.files && imgRef.current.files[0]) {
-			console.log(imgRef.current?.files);
 			const file = imgRef.current.files[0];
 			const webpfile = await webpImageIncoder(file);
 			setImgFile(webpfile + ''); // unknown type을 string으로 변환
